initial commit 🎉
for_any is a library that allows to simplify the process of executing the same function for any variant of an enum. The main goal is to reduce boilerplate in cases where the user has an enum with a bunch of variants, where the types of each variant implement a trait. Traditionally there's no other way then a huge match statement. With for_any, the code would look like `enum_instance.for_any_ref(|field: &dyn Trait| field.trait())`.